+++ src/c++/lib/manta/test/SVLocusScannerTest.cpp
@@ -334,10 +334,9 @@ BOOST_AUTO_TEST_CASE( test_SemiAlignedReads )
 
 BOOST_AUTO_TEST_CASE( test_getSVCandidatesFromReadIndels )
 {
-    const bool isRNA(false);
     const bool isStranded(true);
     const ReadScannerOptions opt;
-    const ReadScannerDerivOptions dopt(opt,isRNA,isStranded);
+    const ReadScannerDerivOptions dopt(opt,isStranded);
 
     ALIGNPATH::path_t inputPath;
     cigar_to_apath("100M2000D100M",inputPath);